+++ lib/rules/object.js
@@ -59,7 +59,7 @@ module.exports = function(schema, path, messages, context) {
 			const newPath = (path ? path + "." : "") + property;
 			//const valueName = `value_${cnt}`;
 
-			src.push(`\n// Field: ${newPath}`);
+			src.push(`\n// Field: ${escapeEvalString(newPath)}`);
 			src.push("if (true) {");
 			src.push(`\tlet field = parentField ? parentField + "${safeSubName}" : "${name}";`);
 			src.push(`\tlet value = ${safePropName};`);
@@ -78,7 +78,7 @@ module.exports = function(schema, path, messages, context) {
 				// Required field
 				src.push(`
 				if (value === undefined || value === null) {
-					${this.makeError({ type: "required", field: newPath, actual: "origValue", messages: rule.messages })}
+					${this.makeError({ type: "required", actual: "origValue", messages: rule.messages })}
 				} else {`);
 			}
