+++ lib/validator.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const flatten = require("./helpers/flatten");
+//const flatten = require("./helpers/flatten");
 const deepExtend = require("./helpers/deep-extend");
 const defaultMessages = require("./messages");
 
@@ -65,8 +65,8 @@ Validator.prototype.validate = function(obj, schema) {
  * @throws {Error} Invalid schema
  */
 Validator.prototype.compile = function(schema) {
-	if (schema === null || typeof schema !== "object" || Array.isArray(schema)) {
-		throw new Error("Invalid schema!");
+	if (schema === null || typeof schema !== "object") {
+		throw new Error("Invalid schema.");
 	}
 
 	const self = this;
@@ -78,16 +78,17 @@ Validator.prototype.compile = function(schema) {
 
 	if (schema.$$root !== true) {
 		if (Array.isArray(schema)) {
-			schema = {
-				type: "multi",
-				rules: schema
-			};
+			const rule = this.compileSchemaRule(schema);
+			schema = rule.schema;
 		} else {
+			const prevSchema = Object.assign({}, schema);
 			schema = {
 				type: "object",
 				strict: schema.$$strict,
-				properties: schema
+				properties: prevSchema
 			};
+
+			delete prevSchema.$$strict;
 		}
 	}
 
@@ -142,8 +143,8 @@ Validator.prototype.compile = function(schema) {
 	const src = sourceCode.join("\n");
 
 	if (this.opts.debug) {
-		//console.log(src);
-		console.log(require("./helpers/prettier")(src));
+		console.log(src);
+		//console.log(require("./helpers/prettier")(src));
 	}
 
 	const checkFn = Function("data", "context", src);
@@ -159,6 +160,9 @@ Validator.prototype.compileSchemaRule = function(schema) {
 			type: schema
 		};
 	} else if (Array.isArray(schema)) {
+		if (schema.length == 0)
+			throw new Error("Invalid schema.");
+
 		schema = {
 			type: "multi",
 			rules: schema
@@ -167,7 +171,7 @@ Validator.prototype.compileSchemaRule = function(schema) {
 
 	const ruleFunction = this.rules[schema.type];
 	if (!ruleFunction)
-		throw new Error("Invalid '" + schema.type + "' type in validator schema!");
+		throw new Error("Invalid '" + schema.type + "' type in validator schema.");
 
 	return {
 		messages: Object.assign({}, this.messages, schema.messages),
@@ -191,37 +195,6 @@ Validator.prototype.makeError = function({ type, field, expected, actual, messag
 		.join(", ");
 
 	return `errors.push({ ${s} });`;
-};
-
-/**
- * Handle results from validator functions
- *
- * @param {Array} errors
- * @param {String} fieldPath
- * @param {Array|Object} res
- */
-Validator.prototype.humanizeErrors = function(errors) {
-	return errors.map(err => {
-		if (!err.message)
-			err.message = this.resolveMessage(err, this.opts.messages[err.type]);
-
-		return err;
-	});
-};
-
-
-/**
- * Resolve message string from a validation error object
- *
- * @param {Object} err Validation error object
- */
-Validator.prototype.resolveMessage = function(err, msg = null) {
-
-	if (msg != null) {
-		const expected = err.expected != null ? err.expected : "";
-		const actual = err.actual != null ? err.actual : "";
-		return msg.replace(/\{field\}/g, err.field).replace(/\{expected\}/g, expected).replace(/\{actual\}/g, actual);
-	}
 };
 
 /**
